+++ src/tools/sceneeditor/EditorRenderer.hpp
@@ -25,7 +25,7 @@ public:
 	void setScene(const std::shared_ptr<Scene> & scene);
 
 	/** \copydoc Renderer::draw */
-	void draw(const Camera & camera, Framebuffer & framebuffer, size_t layer = 0) override;
+	void draw(const Camera & camera, Framebuffer & framebuffer, uint layer = 0) override;
 
 private:
